Abstracts method tree view into a separate component
