+++ ui/app/components/MethodTreeComponent/MethodTreeComponent.css
@@ -1,0 +1,104 @@
+.label {
+  color: #666;
+}
+
+.bold {
+  font-weight: bold;
+}
+
+.filter {
+  padding: 6px;
+  border: 1px solid #eee;
+  margin-left: 20px;
+  font-size: 0.7em;
+}
+
+.alert {
+  padding: 20px;
+  background: #fff;
+  color: #ef5350;
+  font-weight: bold;
+  font-size: 16px;
+}
+
+.card {
+  padding: 4px;
+  box-shadow: rgba(0, 0, 0, 0.117647) 0px 1px 6px, rgba(0, 0, 0, 0.117647) 0px 1px 4px;
+  border-radius: 2px;
+}
+
+.listItem {
+  display: inline-block;
+  height: 14px;
+}
+
+.hover {
+  padding: 1px;
+  border-bottom: 1px solid rgba(158, 158, 158, 0.28);
+  &:HOVER {
+    background: #eee;
+  }
+}
+
+.pill {
+  display: inline-block;
+  width: 90px;
+  font-weight: bold;
+  color: #ff4081;
+}
+
+.onCPU {
+  position: absolute;
+  right: 110px;
+  width: 100px;
+}
+
+.onStack {
+  position: absolute;
+  right: 0px;
+  width: 100px;
+}
+
+.relative {
+  position: relative;
+}
+
+.code {
+  position: absolute;
+  text-overflow: ellipsis;
+  overflow: hidden;
+  left: 14px;
+  white-space: nowrap;
+  right: 220px;
+}
+
+.heading {
+  font-weight: bold;
+  text-align: center;
+}
+
+.number {
+  text-align: left;
+  display: inline-block;
+  width: 30%;
+}
+
+.percentage {
+  text-align: center;
+  display: inline-block;
+  position: relative;
+  width: 70%;
+  .shade {
+    position: absolute;
+    left: 0;
+    top: 0;
+    bottom: 0;
+    background: #ff4081;
+    opacity: 0.2;
+    height: inherit;
+  }
+}
+
+.yellow {
+  background: #9DFF0A;
+}
