+++ ui/app/components/MethodTreeComponent/index.js
@@ -1,0 +1,3 @@
+import MethodTreeComponent from './MethodTreeComponent';
+
+export default MethodTreeComponent;
